Test video modal fade on econimpact
