+++ src/components/blocks/economic-impact/human-impact-banky.js
@@ -23,6 +23,38 @@ const QuoteSource = styled.p`
     color: var(--uog-blue);
 `
 
+
+const Modal=({children}) => (
+    <div id="bankyVideoModal" className="modal fade" tabindex="-1">
+        <div className="modal-dialog modal-lg">
+            <div className="modal-content">
+                <div className="modal-body p-0">
+                    {children}
+                </div>
+            </div>
+        </div>
+    </div>
+)
+
+const Video = ({src, transcript, captions}) => (
+    <div id="video-embed-banky" className="bg-dark">
+        <section name="vimeo" className="ui-kit-section">
+            <div className="embed-responsive embed-responsive-16by9">
+                <div className="d-flex justify-content-center">
+                    <div className="spinner-grow text-yellow" role="status">
+                        <span className="sr-only">Loading...</span>
+                    </div>
+                </div>
+                <video className="ugplayer embed-responsive-item" width="100%" preload="none" controls="controls">
+                    <source type="video/vimeo" src={`${src}`} />
+                    { captions && <track className="caption-input" label="English" kind="captions" srclang="en" src={captions} default="true" /> }
+                    { transcript && <link className="transcript-input" rel="transcript" label="English" kind="descriptions" srclang="en" src={transcript} default="true" /> }
+                </video>
+            </div>
+        </section>
+    </div>
+)
+
 const render = ({ title, body, images, video, testimonial }) => (
     <div className="d-flex flex-column bg-dark">
         <Overlay.GatsbyImage gatsbyImageData={getImage(images.background.src)} alt={images.background.alt}>
@@ -31,7 +63,16 @@ const render = ({ title, body, images, video, testimonial }) => (
                     <Col lg={6} className="fs-3 mb-4">
                         <SectionTitle>{title}</SectionTitle>
                         {body.map((paragraph, index) => <p key={`banky-text-${index}`}>{paragraph}</p>)}
-                        <a className="btn btn-primary my-4" href={video.url}><i class="fa-solid fa-play"></i> Watch Video<span className="visually-hidden">: {video.title}</span></a>
+                        {/* <a className="btn btn-primary my-4" href={video.url}>
+                            <i className="fa-solid fa-play"></i> Watch Video<span className="visually-hidden">: {video.title}</span>
+                        </a> */}
+
+                        <button type="button" className="btn btn-primary my-4" data-bs-toggle="modal" data-bs-target="#bankyVideoModal">
+                            <i className="fa-solid fa-play"></i> Watch Video<span className="visually-hidden">: {video.title}</span>
+                        </button>
+                        <Modal>
+                            <Video src={video.url} transcript={video.transcript} captions={video.captions} />
+                        </Modal>
                     </Col>
                     <Col lg={6} className="d-flex justify-content-center">
                         <GatsbyImage image={getImage(images.foreground.src)} alt={images.foreground.alt} className="align-self-end img-fluid" />
@@ -79,6 +120,8 @@ const query = graphql`
         video {
             title
             url
+            transcript
+            captions
         }
         testimonial {
             quote
