--- src/components/blocks/economic-impact/human-impact-videotest.js
@@ -1,41 +1,0 @@
-import React from "react"
-import { StaticQuery, graphql } from "gatsby"
-import { Container, Row, Col } from "react-bootstrap"
-import styled from "styled-components"
-import EconImpactHumanImpactStats from "components/blocks/economic-impact/human-impact-stats"
-import EconImpactHumanImpactBankyVideoTest from "components/blocks/economic-impact/human-impact-banky-videotest"
-
-const Lead = styled.p`
-  color: #000;
-`
-
-const render = ({ title, lead, body }) => (
-  <>
-    <Container className="page-container mt-5">
-        <Row className="site-content">
-          <h2 id="human-impact">{title}</h2>
-          <Col lg={7} className="mb-5">
-              <Lead className="text-uppercase"><strong className="mt-0">{lead}</strong></Lead>
-              {body.map((paragraph, index) => <p key={`human-impact-text-${index}`}>{paragraph}</p>)}
-          </Col>
-          <EconImpactHumanImpactStats />
-        </Row>
-    </Container>
-    <EconImpactHumanImpactBankyVideoTest />
-  </>
-)
-
-const query = graphql`
-  query {
-    economicImpactYaml(yamlId: {eq: "economic_impact_human_impact"}) {
-      id
-      title
-      lead
-      body
-    }
-  }
-`
-
-export default function EconImpactHumanImpactVideoTest () {
-  return <StaticQuery query={query} render={({economicImpactYaml}) => render(economicImpactYaml)} />
-}
